avatar router updated & audit fix --force
