+++ package.json
@@ -24,7 +24,7 @@
   },
   "devDependencies": {
     "env-cmd": "^10.1.0",
-    "nodemon": "^2.0.20"
+    "nodemon": "^3.0.1"
   },
   "name": "devstore-backend",
   "version": "1.0.0",
